refactor(landing): 메인 랜딩 페이지 폴더 구조 수정
/src/entities/featureCard  -> featureCard 컴포넌트 1개로 분리
/src/entities/featureCard/assets -> eatureCard 에서만 사용되는 이미지 폴더

/src/entities/plansCard  -> plansCard 컴포넌트 1개로 분리

/src/widgets/landing/assets -> landing 페이지 내에서만 사용되는 이미지 폴더
