+++ src/entities/featureCard/assets/featureCard.jpg
