+++ src/entities/featureCard/index.ts
@@ -1,0 +1,1 @@
+import { FeatureCard } from './FeatureCard'
