+++ src/entities/plansCard/config/plansCardItem.ts
