+++ src/entities/plansCard/config/types.ts
@@ -1,0 +1,7 @@
+export interface PlansCardItem {
+  planName: string
+  price: string
+  period: string | null
+  features: { label: string; active: boolean }[]
+  buttonLabel: string
+}
