+++ src/entities/plansCard/index.ts
@@ -1,0 +1,1 @@
+import { PlansCard } from './PlansCard'
