+++ src/pages/home/ui/HomePage.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useEffect } from 'react'
-import { Landing } from '@/widgets/landing'
+import { FeatureSection, HeroMain, PlansSection } from '@/widgets/landing'
 
 export default function HomePage() {
   useEffect(() => {
@@ -16,7 +16,15 @@ export default function HomePage() {
 
   return (
     <>
-      <Landing />
+      <HeroMain />
+      <div className='snap-start'>
+        <section className='grid grid-cols-1 gap-(--spacing-md) px-[var(--spacing-md)] py-[var(--spacing-56)] md:grid-cols-2 lg:grid-cols-3'>
+          <FeatureSection />
+        </section>
+        <section className='px-[var(--spacing-md)] py-[var(--spacing-56)]'>
+          <PlansSection />
+        </section>
+      </div>
     </>
   )
 }
