+++ src/widgets/landing/index.ts
@@ -1,1 +1,3 @@
-export { Landing } from './Landing'
+export { FeatureSection } from './ui/FeatureSection'
+export { HeroMain } from './ui/HeroMain'
+export { PlansSection } from './ui/PlansSection'
